# spec/openapi.yaml
openapi: 3.0.3
info:
  title: blurd
  version: "0.12.0"
  description: >
    Machine-to-machine image redaction. Submit an image by URL or as raw bytes;
    blurd detects faces and licence plates, redacts them, stores only the
    redacted result, and links it to the sha256 of the source.
    This document is the contract every blurd implementation must satisfy.
servers:
  - url: http://127.0.0.1:8770
security:
  - bearerAuth: []
# A key may be SCOPED (see spec/scoped-keys.md). A scope restricts every read
# and stamps every write, and it defines a tenant that namespaces unique codes
# and owns labels. Out-of-scope reads return 404 rather than 403, so a scoped
# caller cannot probe for what exists on the instance.
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer }
  schemas:
    Error:
      type: object
      properties:
        ok: { type: boolean, enum: [false] }
        error:
          type: object
          properties:
            code: { type: integer, description: "Semantic exit code, mirrored by the CLI" }
            type: { type: string }
            message: { type: string }
            details: { type: object }
            recoverable: { type: boolean }
            retry_after: { type: integer, nullable: true }
            suggestions: { type: array, items: { type: string } }
    Job:
      type: object
      description: >
        What a producer gets back from POST /v1/images. Created immediately;
        `result` is populated once status is "done".
      properties:
        job_id: { type: string, description: "job_<16 hex>" }
        status: { type: string, enum: [queued, running, done, failed] }
        external_id: { type: string, nullable: true, description: "the producer's own unique code" }
        source_sha: { type: string, nullable: true, description: "known only once processed" }
        profile_hash: { type: string }
        source_kind: { type: string, enum: [url, stream, code-hit] }
        cached: { type: boolean, description: "true when nothing was reprocessed" }
        attempts: { type: integer }
        created_at: { type: string }
        started_at: { type: string, nullable: true }
        finished_at: { type: string, nullable: true }
        duration_ms: { type: number, nullable: true }
        error: { type: object, nullable: true, description: "same shape as Error.error" }
        result: { $ref: '#/components/schemas/Artifact' }
        blob_url: { type: string }
        blob_url_by_code: { type: string }
    Artifact:
      type: object
      properties:
        source_sha: { type: string, description: "sha256 of the source bytes" }
        profile_hash: { type: string }
        cached: { type: boolean, description: "true when nothing was reprocessed" }
        needs_review: { type: boolean }
        blob:
          type: object
          properties:
            url: { type: string }
            sha256: { type: string }
            bytes: { type: integer }
            mime: { type: string }
        image:
          type: object
          properties:
            width: { type: integer }
            height: { type: integer }
            mime: { type: string }
            source_kind: { type: string, enum: [url, file, stream] }
            source_ref: { type: string, nullable: true }
        codes: { type: array, items: { type: string }, description: "external ids bound to this source, within the caller's tenant" }
        codes_detail:
          type: array
          description: "unscoped callers only: which tenant owns each code"
          items:
            type: object
            properties:
              tenant: { type: string }
              external_id: { type: string }
        tags: { type: array, items: { type: string } }
        metadata: { type: object, additionalProperties: { type: string } }
        detections:
          type: array
          items:
            type: object
            properties:
              cls: { type: string, enum: [face, plate] }
              box: { type: array, items: { type: integer }, minItems: 4, maxItems: 4 }
              score: { type: number }
              detector: { type: string }
        stats:
          type: object
          properties:
            timings:
              type: object
              description: "Per-stage milliseconds: fetch, decode, detect_face, detect_plate, redact, encode, store, total"
            detections: { type: object }
            models: { type: object }
            redact: { type: string }
        created_at: { type: string }
        errors: { type: array, items: { type: string } }
        warnings: { type: array, items: { type: string } }
paths:
...
  /v1/images:
    post:
      summary: Submit an image for redaction (asynchronous)
      description: >
        Always creates a job and returns 202 immediately -- an HTTP request must
        not have to stay open while an image is fetched and processed. Pass
        ?wait=N to block for up to N seconds; if the job settles within that
        window the response is 200 with `result` filled in.
        If `external_id` is already known and on_conflict=reuse (the default),
        the job is returned already "done" without any fetch or decode.
      parameters:
        - { in: query, name: code, schema: { type: string }, description: "external_id; raw-body form. Often the filename." }
        - { in: query, name: wait, schema: { type: number }, description: "seconds to block, max 120. 0 = return at once." }
        - { in: query, name: on_conflict, schema: { type: string, enum: [reuse, replace, reject], default: reuse } }
        - { in: query, name: tags, schema: { type: string }, description: "comma separated (raw-body form)" }
        - { in: query, name: metadata, schema: { type: string }, description: "JSON object (raw-body form)" }
        - { in: query, name: profile, schema: { type: string }, description: "JSON profile overrides" }
        - { in: query, name: force, schema: { type: string }, description: "1 to bypass the cache" }
      requestBody:
        content:
          image/jpeg: { schema: { type: string, format: binary } }
          image/png: { schema: { type: string, format: binary } }
          application/octet-stream: { schema: { type: string, format: binary } }
          application/json:
            schema:
              type: object
              required: [url]
              properties:
                url: { type: string }
                external_id: { type: string }
                on_conflict: { type: string, enum: [reuse, replace, reject] }
                tags: { type: array, items: { type: string } }
                metadata: { type: object }
                profile: { type: object }
      responses:
        "202": { description: "Job accepted; Location header points at /v1/jobs/<id>", content: { application/json: { schema: { $ref: '#/components/schemas/Job' } } } }
        "200": { description: "Job already settled (cache hit, or ?wait covered it)", content: { application/json: { schema: { $ref: '#/components/schemas/Job' } } } }
        "401": { description: "Missing or invalid API key", content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        "422": { description: "Validation failed (bad image, non-public URL, too large)" }
    get:
      summary: List stored artifacts
      parameters:
        - { in: query, name: tag, schema: { type: array, items: { type: string } }, description: "repeatable; ANDed" }
        - { in: query, name: "meta.<key>", schema: { type: string }, description: "exact metadata match" }
        - { in: query, name: sha, schema: { type: string }, description: "sha256 prefix" }
        - { in: query, name: code, schema: { type: string }, description: "external id; exact, or a trailing * for a prefix scan" }
        - { in: query, name: needs_review, schema: { type: string } }
        - { in: query, name: since, schema: { type: string } }
        - { in: query, name: until, schema: { type: string } }
        - { in: query, name: sort, schema: { type: string, enum: [created, faces, plates, size, review], default: created } }
        - { in: query, name: direction, schema: { type: string, enum: [desc, asc], default: desc } }
        - { in: query, name: cursor, schema: { type: string }, description: "next_cursor from the previous page; keyset pagination, O(1) at any depth" }
        - { in: query, name: limit, schema: { type: integer, default: 50, maximum: 200 } }
        - { in: query, name: offset, schema: { type: integer, default: 0 }, description: "legacy; prefer cursor -- OFFSET is O(offset) and unstable under concurrent inserts" }
      responses:
        "200":
          description: >
            total, total_capped, count, items[], next_cursor, has_more, sort,
            direction. `total` stops counting at 10000 (total_capped=true) and
            is returned only when no cursor was supplied, because paging deeper
            cannot change it.
...
  /v1/jobs:
    get:
      summary: List jobs
      parameters:
        - { in: query, name: status, schema: { type: string, enum: [queued, running, done, failed] } }
        - { in: query, name: code, schema: { type: string } }
        - { in: query, name: since, schema: { type: string } }
        - { in: query, name: until, schema: { type: string } }
        - { in: query, name: sort, schema: { type: string, enum: [created, duration, status], default: created } }
        - { in: query, name: direction, schema: { type: string, enum: [desc, asc], default: desc } }
        - { in: query, name: cursor, schema: { type: string } }
        - { in: query, name: limit, schema: { type: integer, default: 50, maximum: 200 } }
        - { in: query, name: offset, schema: { type: integer, default: 0 } }
      responses:
        "200":
          description: >
            total, total_capped, count, items[], next_cursor, has_more, queue{}.
            Items in a LISTING omit `result`; fetch a single job to get it.
...